data_crw/Scraper0.1/requirements.txt
# Python package dependencies for Facebook scraper
# Core dependencies
selenium>=4.15.0
webdriver-manager>=4.0.0
sqlalchemy>=2.0.20
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0

# Data processing and utilities
requests>=2.31.0
pandas>=2.1.0
numpy>=1.26.0
tqdm>=4.66.0
python-dateutil>=2.9.0
dateparser>=1.2.0

# Web scraping and parsing
beautifulsoup4>=4.12.2
lxml>=5.0.0

# Image processing
pillow>=10.0.0

# Command line interface
tabulate>=0.9.0
colorama>=0.4.6

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Machine Learning
scikit-learn>=1.3.0
# catboost>=1.2  # Disabled - may have compatibility issues
joblib>=1.3.1

# Firebase
firebase-admin>=6.2.0

# FastAPI
fastapi>=0.109.0
uvicorn>=0.25.0
pydantic>=2.5.0

# Google Sheets
gspread>=5.12.0
oauth2client>=4.1.3

# NLP for sentiment analysis
textblob>=0.17.1
nltk>=3.8.1
